Adds new metrics assertion util file
